+++ src/main/java/com/dwarfeng/dcti/sdk/util/DataInfoUtil.java
@@ -73,6 +73,24 @@ public class DataInfoUtil {
         dataInfo.setHappenedDateNanoOffset(TimeUtil.toNanoOffset(instant));
     }
 
+    /**
+     * 根据指定参数构造新的数据信息。
+     *
+     * @param pointLongId     指定的数据点 ID。
+     * @param value           指定的数据值。
+     * @param happenedInstant 指定的发生时间点。
+     * @return 构造的数据信息。
+     * @throws NullPointerException 参数 <code>happenedInstant</code> 为 <code>null</code>。
+     * @since 2.0.1
+     */
+    @Nonnull
+    public static DataInfo newInstance(long pointLongId, String value, @Nonnull Instant happenedInstant) {
+        Objects.requireNonNull(happenedInstant, "happenedInstant 不能为空");
+        return new DataInfo(
+                pointLongId, value, TimeUtil.toDate(happenedInstant), TimeUtil.toNanoOffset(happenedInstant)
+        );
+    }
+
     private DataInfoUtil() {
         throw new IllegalStateException("禁止外部实例化");
     }