+++ src/test/java/com/dwarfeng/dcti/sdk/util/TimedValueUtilTest.java
@@ -71,4 +71,22 @@ public class TimedValueUtilTest {
         timedValue.setHappenedDateNanoOffset(TimeUtil.MIN_NANO_OFFSET - 1);
         TimedValueUtil.getHappenedInstant(timedValue);
     }
+
+    @Test
+    public void testNewInstance() {
+        String expectedValue = "hello";
+        Instant expectedInstant = Instant.ofEpochSecond(1713333333L, 123456789L);
+
+        TimedValue timedValue = TimedValueUtil.newInstance(expectedValue, expectedInstant);
+
+        Assert.assertEquals(expectedValue, timedValue.getValue());
+        Assert.assertEquals(expectedInstant, TimedValueUtil.getHappenedInstant(timedValue));
+    }
+
+    // 该方法本身是用来进行异常测试的，故忽略相关异常警告。
+    @SuppressWarnings("DataFlowIssue")
+    @Test(expected = NullPointerException.class)
+    public void testNewInstanceWithNullInstant() {
+        TimedValueUtil.newInstance("hello", null);
+    }
 }